Add tests for kernel prover.
